# turerp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Turerp ERP API",
    "description": "Multi-tenant SaaS ERP system API\n\n## Authentication\n\nAll endpoints except `/api/v1/auth/login`, `/api/v1/auth/register`, and `/api/v1/auth/refresh` require JWT Bearer token authentication.\n\n## Rate Limiting\n\nAuthentication endpoints are rate limited to 10 requests per minute per IP address with a burst of 3 requests.\n\n## Localization\n\nAll endpoints support per-request localization via the `Accept-Language` header. Supported languages: `en` (default), `tr`.\n\n## API Versioning\n\n- `/api/v1/` - Current stable API (recommended)\n- `/api/auth/`, `/api/users/` - Legacy routes (deprecated, will be removed in v2)",
    "contact": {
      "name": "Turerp Team",
      "email": "info@turerp.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/feature-flags": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Get all feature flags endpoint (authenticated)",
        "operationId": "get_flags",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "Page number (1-based, default: 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "description": "Number of items per page (default: 20, max: 100)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flags retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResult"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Create feature flag endpoint (admin only)",
        "operationId": "create_flag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeatureFlag"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Feature flag created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "409": {
            "description": "Feature flag already exists"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/feature-flags/check/{name}": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Check if feature is enabled endpoint (authenticated)",
        "operationId": "check_feature",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Feature flag name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/feature-flags/{id}": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Get feature flag by ID endpoint (authenticated)",
        "operationId": "get_flag_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Feature flag ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "404": {
            "description": "Feature flag not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Update feature flag endpoint (admin only)",
        "operationId": "update_flag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Feature flag ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeatureFlag"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Feature flag updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Feature flag not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Delete feature flag endpoint (admin only)",
        "operationId": "delete_flag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Feature flag ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Feature flag not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/feature-flags/{id}/disable": {
      "post": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Disable feature flag endpoint (admin only)",
        "operationId": "disable_flag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Feature flag ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Feature flag not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/feature-flags/{id}/enable": {
      "post": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "Enable feature flag endpoint (admin only)",
        "operationId": "enable_flag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Feature flag ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated - missing or invalid JWT token"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Feature flag not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateFeatureFlag": {
        "type": "object",
        "description": "Create feature flag request",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Human-readable description (optional, max 500 characters)",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Feature flag name (required, 1-100 characters, alphanumeric with underscores/hyphens)"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FeatureFlagStatus"
              }
            ],
            "nullable": true
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64",
            "description": "Optional tenant ID for tenant-specific flags",
            "nullable": true
          }
        }
      },
...
      "FeatureFlagResponse": {
        "type": "object",
        "description": "Feature flag response for API",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/chrono.NaiveDateTime"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FeatureFlagStatus"
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "updated_at": {
            "$ref": "#/components/schemas/chrono.NaiveDateTime"
          }
        }
      },
      "FeatureFlagStatus": {
        "type": "string",
        "description": "Feature flag status",
        "enum": [
          "enabled",
          "disabled"
        ]
      },
      "FeatureStatusResponse": {
        "type": "object",
        "description": "Response for feature status check",
        "required": [
          "name",
          "enabled",
          "tenant_id"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Whether the feature is enabled"
          },
          "name": {
            "type": "string",
            "description": "Feature flag name"
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64",
            "description": "Tenant ID context"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "description": "Simple localized success message payload.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateFeatureFlag": {
        "type": "object",
        "description": "Update feature flag request",
        "properties": {
          "description": {
            "type": "string",
            "description": "Human-readable description (optional, max 500 characters)",
            "nullable": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FeatureFlagStatus"
              }
            ],
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}